+++ sewar/no_ref.py
@@ -41,7 +41,7 @@ def d_s (pan,ms,fused,q=1,r=4,ws=7):
 	pan = pan.astype(np.float64)
 	fused = fused.astype(np.float64)
 	
-	pan_degraded = uniform_filter(pan, size=ws)/(ws**2)
+	pan_degraded = uniform_filter(pan.astype(np.float64), size=ws)/(ws**2)
 	pan_degraded = imresize(pan_degraded,(pan.shape[0]//r,pan.shape[1]//r))
 
 	L = ms.shape[2]
